Rift: detail position report
As documented in the DK2 Firmware Specification, reserved_1 actually is
a version field to indicate whether the values are hard-coded defaults,
factory calibrated, or user calibrated.

reserved_2 is a rotation angle around the direction vector, which is not
actually used on Rift DK2 and CV1, presumably because IMU rotation can
be determined from the IMU calibration report.