REUSE 3.0 compliance
Update copyright and license headers for compliance with version 3.0 of
the REUSE specification. Drop the AUTHORS file, there's a copyright header
in every file now.

This replaces all GPL-2.0+ with GPL-2.0-or-later license identifiers.
All LGPL-2.0+ license identifiers are replaced with LGPL-2.1-or-later.
The composite expressions are encapsulated by parentheses and the "OR"
operator is written in uppercase as the SPDX Specification 2.1 suggests.